Eliminate a data copy in the recv path
The recv logic was marshaling data into an ei buffer and then, if too large
to fit in the fixed-size outgoing buffer going back to Erlang, copied into
a newly-allocated binary. Change the flow to allocate the binary when we
know the data is too large for the fixed-size buffer, and marshal the term
directly into that binary.