Use rlimit.rlim_cur for socket tracking table size
On OS X Yosemite, getrlimit can return a large bogus value for the
rlimit.rlim_max field. This field was used during enm driver
initialization as part of allocating a file descriptor table, and the
bogus value was causing a memory allocation failure with driver_alloc
because the size being requested was much too high. Use the
rlimit.rlim_cur field instead to calculate the amount of memory needed
for the table.