Fix #15: return after driver_failure calls
In the driver, functions calling driver_failure were not returning
immediately afterward, and so any code in the remainder of such a
function that tried to access driver state could cause access
violations due to that state already having been destroyed. Fix these
calls, and also modify which driver_failure functions are called so as
to make errors as descriptive as possible.